EggBreak Script
I was trying to make a script so that when the 'Player tag' object (the controllers, I was testing with a block the egg was rolling into) hits the egg collider it will play the little pulse animation. The ideally at 3rd hit it would delete the main 'complete egg' and trigger rigidbodies and sphere colliders on the shell portions to have them split apart, and then spit out the animal.